+++ rana_qgis_plugin/widgets/utils_file_action.py
@@ -2,6 +2,7 @@ from enum import Enum
 from typing import List, Optional
 
 from qgis.PyQt.QtCore import QObject, pyqtSignal
+from qgis.PyQt.QtWidgets import QApplication
 
 from rana_qgis_plugin.auth_3di import has_3di_authcfg
 from rana_qgis_plugin.constant import SUPPORTED_DATA_TYPES
@@ -9,11 +10,13 @@ from rana_qgis_plugin.utils.api import FileDescriptorStatus, get_tenant_file_des
 
 
 class FileAction(Enum):
-    # Actions related to downloading and local files
+    # Actions related to accessing data
     OPEN_IN_QGIS = "Open in QGIS"
     OPEN_WMS = "Open WMS in QGIS"
     DOWNLOAD_RESULTS = "Download Results"
     OPEN_IN_FILE_BROWSER = "Open in file browser"
+    OPEN_IN_BROWSER = "Open in browser"
+    COPY_WMS_URL = "Copy WMS URL"
     # Actions related to viewing or modifying files on Rana
     SAVE_REVISION = "Upload to Rana"
     # Saving vector and raster styling follows a different path and thus there are different actions
@@ -22,11 +25,12 @@ class FileAction(Enum):
     # SAVE_STYLING action is not used atm, only defined for naming
     SAVE_STYLING = "Save style to Rana"
     UPLOAD_FILE = "Save Data to Rana"
+    EXPORT_GPKG = "Export to GeoPackage"
     VIEW_REVISIONS = "View all Revisions"
+    HISTORY = "History"
     RENAME = "Rename"
     DELETE = "Delete"
     REMOVE_FROM_PROJECT = "Remove from Project"
-    OPEN_IN_BROWSER = "Open in browser"
 
     def __lt__(self, other):
         # sort a list of file actions by order of definition here
@@ -71,7 +75,7 @@ def get_file_actions_by_data_type(data_type: str) -> List[FileAction]:
         actions = [FileAction.REMOVE_FROM_PROJECT] + actions[1:]
         if has_3di_authcfg():
             actions += [FileAction.SAVE_REVISION, FileAction.VIEW_REVISIONS]
-        actions += [FileAction.OPEN_IN_BROWSER]
+        actions += [FileAction.EXPORT_GPKG, FileAction.OPEN_IN_BROWSER]
     return sorted(actions)
 
 
@@ -83,6 +87,7 @@ def get_scenario_actions(
         actions.append(FileAction.DOWNLOAD_RESULTS)
         if meta["simulation"]["software"]["id"] == "3Di":
             actions.append(FileAction.OPEN_WMS)
+            actions.append(FileAction.COPY_WMS_URL)
         # Add open in file browser to any file type that can be opened
         # Actual check for file availibility will be done downstream
         actions.append(FileAction.OPEN_IN_FILE_BROWSER)
@@ -95,6 +100,26 @@ def get_scenario_actions(
     return actions
 
 
+def copy_wms_url_to_clipboard(file: dict, communication=None):
+    """Copy the WMS URL of a file to the clipboard.
+
+    Args:
+        file: The file dict containing a descriptor_id.
+        communication: Optional UICommunication instance for user feedback.
+    """
+    descriptor = get_tenant_file_descriptor(file["descriptor_id"])
+    wms_link = next(
+        (link for link in descriptor["links"] if link["rel"] == "wms"), None
+    )
+    communication.log_info(f"WMS URL: {wms_link['href']}")
+    if wms_link:
+        QApplication.clipboard().setText(wms_link["href"])
+        if communication:
+            communication.bar_info("WMS URL copied to clipboard.")
+    elif communication:
+        communication.show_warn("No WMS URL available for this file.")
+
+
 class FileActionSignals(QObject):
     file_deletion_requested = pyqtSignal(dict)
     file_rename_requested = pyqtSignal(dict, str)
@@ -104,6 +129,7 @@ class FileActionSignals(QObject):
     save_raster_styling_requested = pyqtSignal(dict)
     save_revision_requested = pyqtSignal(dict)
     open_wms_requested = pyqtSignal(dict)
+    export_gpkg_requested = pyqtSignal(dict)
     download_file_requested = pyqtSignal(dict)
     download_results_requested = pyqtSignal(dict)
     view_all_revisions_requested = pyqtSignal(dict, dict)
@@ -119,7 +145,9 @@ class FileActionSignals(QObject):
             FileAction.SAVE_RASTER_STYLING: self.save_raster_styling_requested,
             FileAction.SAVE_REVISION: self.save_revision_requested,
             FileAction.OPEN_WMS: self.open_wms_requested,
+            FileAction.EXPORT_GPKG: self.export_gpkg_requested,
             FileAction.DOWNLOAD_RESULTS: self.download_results_requested,
             FileAction.VIEW_REVISIONS: self.view_all_revisions_requested,
+            FileAction.HISTORY: self.view_all_revisions_requested,
         }
         return signal_map.get(signal_type)